Vertical Slice Maps Creation
I created the main map and sub level maps for the work of this week to come. Me, Francesco and Leonardo will work on it from tomorrow to 15/01 but from tomorrow we will push on a new branch created from the develop except for the landscape that can be pushed on the main branch from only one person.